Support user code start at 0x8001800
Introduces a new feature flag ENABLE_FLASH0x8001800 (disabled by default). If this
is enabled, the bootloader supports booting from and writing to 0x8001800, freeing
up 2KB of flash for the user.

Note that for this option to work, others need to be disabled, as the bootloader
won't fit in 6KB otherwise.