Check static data size in linker script
Static data is appended to the binary after the code, but previously its size
wasn't accounted for by the linker when checking flash region limits. Fix this by
explicitly telling the linker to put it in the flash region, and not just after
the code.

If there is enough space and linking succeeds, the output is identical to that
produced with the old section definition.